backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
psycopg2-binary>=2.9.9
django-cors-headers>=4.3.1
requests>=2.32.0
